+++ Assets/Packages/DeviceDebugger/CHANGELOG.md
@@ -1,7 +1,14 @@
 # Changelog
 
+## [1.0.4] - 2026-01-18
+- Update Jenkinsfile
+- Fix [MenuItem("Tools/Device Debugger/Add to Scene")]
+
+## [1.0.3] - 2026-01-09
+- Update Jenkinsfile
+
 ## [1.0.2] - 2025-10-03
- Add [MenuItem("Tools/Device Debugger/Remove from Scene")] 
+- Add [MenuItem("Tools/Device Debugger/Remove from Scene")] 
 
 ## [1.0.1] - 2025-09-25
 - Remove [MenuItem("Tools/Graphy/Create Prefab Variant")]